requirements.txt
tensorflow
tqdm
Pillow
opencv-python
ffmpeg
pyyaml
matplotlib